ai-services/ocr/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0